requirements.txt
openai
python-dotenv
crewai
crewai-tools
langchain_openai